+++ src/Resources/app/storefront/src/smart-payment-buttons/swag-paypal.smart-payment-buttons.js
@@ -142,8 +142,7 @@ export default class SwagPayPalSmartPaymentButtons extends SwagPaypalAbstractBut
     }
 
     createButton() {
-        this.createScript(() => {
-            const paypal = window.paypal;
+        this.createScript((paypal) => {
             this.renderButton(paypal);
         });
     }