+++ src/Resources/app/storefront/src/swag-paypal.script-loading.js
@@ -1,0 +1,9 @@
+/* eslint-disable import/no-unresolved */
+
+export default class SwagPayPalScriptLoading {
+    loadingScript = false;
+
+    paypal = null;
+
+    callbacks = [];
+}