+++ src/Resources/app/administration/src/module/swag-paypal-payment/index.js
@@ -4,7 +4,7 @@ import './component/swag-paypal-payment-details-v1';
 import './component/swag-paypal-payment-details-v2';
 import './component/swag-paypal-text-field';
 import './component/swag-paypal-textarea-field';
-import './extension/sw-order';
+import './extension/sw-order-detail';
 import './page/swag-paypal-payment-detail';
 
 const { Module } = Shopware;