+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # REPLACE_GLOBALLY_WITH_NEXT_VERSION (3.2.1)
 - PPI-279, PPI-297 - Extends the partner referral API
 - PPI-290 - Improved extensibility
+- PPI-296 - Improved display of APMs in Footer
 
 # 3.2.0
 - PPI-262 - Fix issue where Express button was not shown in CMS buy box elements