+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # REPLACE_GLOBALLY_WITH_NEXT_VERSION (3.2.1)
 - PPI-279, PPI-297 - Extends the partner referral API
 - PPI-290 - Improved extensibility
+- PPI-295 - Fix order details in administration not showing complete page
 - PPI-296 - Improved display of APMs in Footer
 - PPI-298 - Fix issues with loading spinner with Smart Payment Buttons